Reduce thread wakeup by replacing the 50ms wait and check with a blocking-queue implementation which allows the JVM to handle the blocking wait internally. This also should reduce unnecessary wakeups of the Java VM considerable and thus reduce power consumption on battery based systems (laptops) as we do not wakeup 20 times per second, but only once per second for shutdown detection.
We could even get rid of this wakeup as well by not using "ExitService.isExiting()" and using a special "ExitEvent" instead, but I didn't want to change this at this time and the gain may not be big anyway.

This also removes the 50ms delay per event which slows down event processing and was noticeable when many events were added at the same time. Now events are processed as fast as possible without delay.

Also remove "volatile" for the queue object, I don't think this is necessary, the member is static and initialized statically, so no synchronization necessary per guarantee by the JVM for static initialization.