Rework TrackComparator for improved performance and reduced memory churn
Performance analysis using Dynatrace showed that this class is allocating
a lot of memory because of all the string concatenation just for comparing
two objects.

The new version now avoids allocations as much as possible and so should cause
much less memory allocation/churn.

In a local microbenchmark also runtime of executing the comparator many times went
from 1100ms down to around 350ms.

Add tests for TrackComparator and make it pass a new general ComparatorTest
Also adjust the base test-classes slightly to verify results of file-operations